perf: land architecture-review punch list (TODO Part 7A/7B/7C)
Addresses the Tier 1/2/3 items documented in TODO.md Part 7, drawn from
the 2026-04-16 architecture review. Motivation is primarily sustained-30fps
headroom and thread-safety hardening; nothing here changes the output wire
format for CIGI, DIS, KLV, MPEG-TS, or CoT.

Tier 1 (perf + correctness):
- Drop per-tick FlushRenderingCommands() from the readback poll path;
  ColorReadbackPool/DepthReadbackPool + bPollComplete_ + PollGeneration_
  atomics let the game thread consume without serialising the render thread.
- Extract Cesium tileset tuning into a shared helper and reapply it from
  HotReloadConfig; disable physics-mesh cook on all tilesets; FOV-scale
  CulledScreenSpaceError for narrow-FOV sensor presets.
- Replace Euler-rate integration in FDRState with a body-frame quaternion
  to kill gimbal-lock discontinuities during steep climb/dive dead-reckon.
- Wrap CIGI/DIS/encoder SPSC queues in TBoundedSpscQueue with drop-newest
  and per-queue drop counters exposed via GetDropCount().

Tier 2 (thread-safety + UE5 idioms):
- Move health JSON + Prometheus file writes off the game thread via
  AsyncTask; cache /metrics body on a TSharedRef<FString> refreshed at ~1 Hz.
- Out-param entity snapshot with capacity reuse + cone cull before AABB
  projection; saves 15k/s projections at MaxEntities=500.
- Async mesh load via FStreamableManager for /Game/ paths; entity stays
  hidden until load completes. Drop const_cast on FEntityTypeTable cache
  by making Set* methods const over mutable maps.
- FRWLock-guarded HotReloadConfig + GetConfigSnapshot() for cross-thread
  readers that need a coherent struct copy.
- Clean CIGI/DIS shutdown via FEvent so recv loops wake immediately on
  Stop(); fail-hard on DIS multicast join failure.
- JSON logger drains on a dedicated consumer thread; full RFC 7159 escape
  (newlines/tabs/control chars/\uXXXX) closes the JSONL-framing gap.

Tier 3 (polish):
- FMath::UnwindDegrees in the gimbal slew wrap.
- Reuse AVPacket + KLV scratch across frames (BuildMisbST0601Into).
- Verbose log on truncated DIS Entity State PDUs for silent-drop diagnosis.
- Reject non-finite / out-of-bounds telemetry in FCotSender::BuildCotXml.
- 64-bit latency tracker WriteIndex so the ring wrap stays exact past the
  4.5-year uint32 horizon.

Verification status: builds clean on Linux + clang 20.1.8. stat unit /
stat GPU before-vs-after and the 29 UE Automation tests still need to run
in a full dev environment (see TODO.md §7D).

+++ unreal_project/CamSimTest/Source/CamSimTest/Diagnostics/PipelineLatencyTracker.cpp
@@ -16,8 +16,8 @@ void FPipelineLatencyTracker::SetStageTimestamp(EPipelineStage Stage, uint64 Cyc
 
 void FPipelineLatencyTracker::CommitFrame()
 {
-	const uint32 Idx = WriteIndex.Load(EMemoryOrder::Relaxed);
-	Records[Idx % BufferCapacity] = CurrentFrame;
+	const uint64 Idx = WriteIndex.Load(EMemoryOrder::Relaxed);
+	Records[static_cast<int32>(Idx % BufferCapacity)] = CurrentFrame;
 	WriteIndex.Store(Idx + 1, EMemoryOrder::SequentiallyConsistent);
 
 	const int32 Count = CommittedCount.Load(EMemoryOrder::Relaxed);
@@ -61,7 +61,7 @@ FPipelineLatencyTracker::FLatencyPercentiles FPipelineLatencyTracker::ComputePer
 	EncodeDeltas.Reserve(Count);
 	TotalDeltas.Reserve(Count);
 
-	const uint32 WIdx = WriteIndex.Load(EMemoryOrder::SequentiallyConsistent);
+	const uint64 WIdx = WriteIndex.Load(EMemoryOrder::SequentiallyConsistent);
 	const int32 StartIdx = (Count < BufferCapacity) ? 0 : static_cast<int32>(WIdx % BufferCapacity);
 
 	for (int32 i = 0; i < Count; ++i)
